refactor coagulator

+++ go/algorithm/coagulator1d/kernel.go
@@ -3,5 +3,5 @@ package coagulator1d
 import "github.com/maxkuzn/advection-and-coagulation/internal/cell"
 
 type Kernel interface {
-	Compute(x, y int) cell.FloatType
+	Compute(x, y float64) cell.FloatType
 }
